Postman tests passed. Download shop_cart to do

+++ backend/foodgramm_backend/api/views.py
@@ -103,9 +103,9 @@ class RecipeViewSet(viewsets.ModelViewSet):
     @action(detail=True, methods=['post', 'delete'],
             permission_classes=[permissions.IsAuthenticated])
     def favorite(self, request, pk):
-        if not Recipe.objects.all().filter(id=pk):
-            return Response('Такого рецепта нет', status=status.HTTP_404_NOT_FOUND)
         if request.method == 'POST':
+            if not Recipe.objects.all().filter(id=pk):
+                return Response('Такого рецепта не существует', status=status.HTTP_400_BAD_REQUEST)
             if Best.objects.filter(user=request.user, recipe__id=pk).exists():
                 return Response('Уже в избранном', status=status.HTTP_400_BAD_REQUEST)
             recipe = get_object_or_404(Recipe, id=pk)
@@ -113,6 +113,8 @@ class RecipeViewSet(viewsets.ModelViewSet):
             serializer = RecipeLimitedSerializer(recipe)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         else:
+            if not Recipe.objects.all().filter(id=pk):
+                return Response('Такого рецепта не существует', status=status.HTTP_404_NOT_FOUND)
             if not Best.objects.filter(user=request.user, recipe__id=pk).exists():
                 return Response('Такого рецепта нет в избранном', status=status.HTTP_400_BAD_REQUEST)
             Best.objects.filter(user=request.user, recipe__id=pk).delete()
@@ -121,9 +123,9 @@ class RecipeViewSet(viewsets.ModelViewSet):
     @action(detail=True, methods=['post', 'delete'],
             permission_classes=[permissions.IsAuthenticated])
     def shopping_cart(self, request, pk):
-        if not Recipe.objects.all().filter(id=pk):
-            return Response('Такого рецепта нет', status=status.HTTP_404_NOT_FOUND)
         if request.method == 'POST':
+            if not Recipe.objects.all().filter(id=pk):
+                return Response('Такого рецепта не существует', status=status.HTTP_400_BAD_REQUEST)
             if ShopCart.objects.filter(user=request.user, recipe__id=pk).exists():
                 return Response('Уже в корзине!', status=status.HTTP_400_BAD_REQUEST)
             recipe = get_object_or_404(Recipe, id=pk)
@@ -131,6 +133,8 @@ class RecipeViewSet(viewsets.ModelViewSet):
             serializer = RecipeLimitedSerializer(recipe)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         else:
+            if not Recipe.objects.all().filter(id=pk):
+                return Response('Такого рецепта не существует', status=status.HTTP_404_NOT_FOUND)
             if not ShopCart.objects.filter(user=request.user, recipe__id=pk).exists():
                 return Response('Такого рецепта нет в корзине', status=status.HTTP_400_BAD_REQUEST)
             ShopCart.objects.filter(user=request.user, recipe__id=pk).delete()
